Modified the file names and methods to include the check for catalog's domain.
